shared components for manual building plans
