+++ app/doctor/plans.tsx
@@ -1,0 +1,5 @@
+import Plans from '../components/AI_Plans/Plans';
+
+export default function DoctorPlansScreen() {
+  return <Plans accountType="doctor" />;
+}
